requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
imageio-ffmpeg
pillow